refector score and finish css
